requirements.txt

Flask==3.0.3
gunicorn==23.0.0
speedtest-cli==2.1.3
